+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 - Support pasted GIFs
+- Group notifications in Notification Center by chat
+- More characters in notification before truncating
+- Fix: separate multiple notifications coming in at the same time
 
 ## v1.58.1
 2025-04